requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
norm==1.1.3
numpy==1.15.4
pandas==0.23.4
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
rope==0.11.0
scipy==1.2.0
six==1.12.0
stopwatch.py==1.0.1
